requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==1.0.0
Django==4.0.6
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
middleware==1.2.3
Pillow==9.1.1
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
twilio==7.12.0
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
